# api/openapi.yaml
openapi: 3.1.0
info:
  title: Skyline MCP Gateway API
  description: >-
    REST and MCP protocol API for the Skyline MCP Gateway server. Manages API
    profiles, tool discovery, tool execution, credential verification, admin
    observability, and OAuth 2.1 flows for ChatGPT MCP compatibility.
  version: 0.9.16
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://localhost:8191
    description: Default local development server
...
security: []
paths:
...
  /admin/stats:
    get:
      operationId: getStats
      summary: Aggregated statistics (audit + metrics)
      tags: [admin]
      security:
        - AdminSession: []
      parameters:
        - name: profile
          in: query
          description: Filter stats by profile name
          schema:
            type: string
        - name: since
          in: query
          description: Start of time window (RFC 3339). Defaults to 24 hours ago.
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Combined audit stats and metrics snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  audit_stats:
                    $ref: '#/components/schemas/AuditStats'
                  metrics_snapshot:
                    $ref: '#/components/schemas/MetricsSnapshot'
                  period:
                    type: object
                    properties:
                      since:
                        type: string
                        format: date-time
                      until:
                        type: string
                        format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
...
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: skyline_admin
      description: >-
        Admin session cookie set by POST /admin/auth. Secure, HttpOnly,
        SameSite=Strict, 7-day expiry.

    ProfileToken:
      type: http
      scheme: bearer
      description: >-
        Per-profile bearer token. Each profile has its own token used for
        authentication when accessing profile data, tools, and MCP endpoints.

  # ──────────────────────────────────────────────
  # Parameters
  # ──────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
...
  schemas:
...
    AuditEvent:
      type: object
      properties:
        id:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
        profile:
          type: string
        event_type:
          type: string
          enum: [execute, connect, disconnect, error]
        api_name:
          type: string
        tool_name:
          type: string
        arguments:
          type: object
          additionalProperties: true
        duration_ms:
          type: integer
          format: int64
        status_code:
          type: integer
        success:
          type: boolean
        error_msg:
          type: string
        client_addr:
          type: string
        request_size:
          type: integer
          format: int64
        response_size:
          type: integer
          format: int64
    AuditStats:
      type: object
      description: Aggregated audit statistics
      properties:
        total_requests:
          type: integer
          format: int64
        successful_requests:
          type: integer
          format: int64
        failed_requests:
          type: integer
          format: int64
        error_rate:
          type: number
          format: double
        avg_duration_ms:
          type: integer
          format: int64
        max_duration_ms:
          type: integer
          format: int64
        min_duration_ms:
          type: integer
          format: int64
        total_request_bytes:
          type: integer
          format: int64
        total_response_bytes:
          type: integer
          format: int64
        est_request_tokens:
          type: integer
          format: int64
        est_response_tokens:
          type: integer
          format: int64
        top_apis:
          type: array
          items:
            $ref: '#/components/schemas/APIStats'
        top_tools:
          type: array
          items:
            $ref: '#/components/schemas/APIStats'
        recent_events:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
    APIStats:
      type: object
      properties:
        name:
          type: string
        calls:
          type: integer
          format: int64
        errors:
          type: integer
          format: int64
        error_rate:
          type: number
          format: double
        avg_ms:
          type: integer
          format: int64
    MetricsSnapshot:
      type: object
      properties:
        total_requests:
          type: integer
          format: int64
        success_requests:
          type: integer
          format: int64
        failed_requests:
          type: integer
          format: int64
        active_connections:
          type: integer
          format: int64
        total_connections:
          type: integer
          format: int64
        avg_duration_ms:
          type: number
          format: double
        cache_hits:
          type: integer
          format: int64
        cache_misses:
          type: integer
          format: int64
        profile_requests:
          type: object
          additionalProperties:
            type: integer
            format: int64
        tool_requests:
          type: object
          additionalProperties:
            type: integer
            format: int64
        uptime_seconds:
          type: number
          format: double
...